+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - deps: update json-as to v1.0.0 and upgrade related code [#799](https://github.com/hypermodeinc/modus/pull/779)
 - fix: bump api explorer to account for recursive types, default required variables with falsy values [#766](https://github.com/hypermodeinc/modus/pull/766)
+- fix: support float32 vector arrays in DQL query variables [#781](https://github.com/hypermodeinc/modus/pull/781)
 
 ## 2025-02-10 - AssemblyScript SDK 0.17.1